Fix race condition.
