+++ app/templates/custom-elements/update-dialog.html
@@ -116,7 +116,6 @@
     getLatestRelease,
     getUpdateStatus,
     getVersion,
-    refreshCsrfToken,
     shutdown,
     update,
   } from "/js/controllers.js";
@@ -290,8 +289,6 @@
 
           try {
             await this._waitForUpdateToFinish();
-            // Updating causes the backend to restart, changing the CSRF token.
-            await refreshCsrfToken();
             await this._performRestart();
             // Wait at least 10 seconds to ensure reboot has started before
             // checking whether the system is back up.
